requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
anthropic>=0.18.0
arxiv>=2.1.0
jinja2>=3.1.0
